requirements.txt
requests
flask
pydantic
groq
python-dotenv
autogen
flaml 
pyyaml
